Disabled XStorage API until ready.
Fixed wrong var type in XboxTick.

+++ xlive/H2MOD/Variants/XboxTick/XboxTick.cpp
@@ -8,7 +8,7 @@ XboxTick::XboxTick()
 void XboxTick::tickRate() {
 
 	DWORD time_globals = *(DWORD*)(h2mod->GetAddress(0x4C06E4, 0x4CF0EC));
-	*(BYTE*)((char*)time_globals + 0x2) = 30;
+	*(short*)((char*)time_globals + 0x2) = 30;
 	*(float*)((char*)time_globals + 0x4) = 0.03333333507f;
 	TRACE_GAME("[h2mod] Set tickrate to 30");
 }
